+++ controllers/monthController.js
@@ -11,23 +11,3 @@ module.exports.get = function(req, res, next) {
         });
     });
 }
-
-module.exports.newTask = function(req, res, next) {
-    storage.storeTask(req.body, getStorageCallback(res, next));
-}
-
-module.exports.editTask = function(req, res, next) {
-    var task = req.body;
-    task.id = req.params.id;
-    storage.editTask(task, getStorageCallback(res, next));
-}
-
-function getStorageCallback(res, next) {
-    return function(err, result) {
-        if (err) {
-            next(err);
-        } else {
-            res.json(result);
-        }
-    }
-}
